requirements.txt
fastapi
strawberry-graphql
uvicorn

